refactor: 인플루언서 관련 리팩터링
